+++ plugins/org.eclipse.osee.framework.core/src/org/eclipse/osee/framework/core/data/RelationalConstants.java
@@ -23,12 +23,12 @@ public final class RelationalConstants {
 
    public static final int JOIN_QUERY_ID_SENTINEL = -1;
    public static final boolean IS_HISTORICAL_DEFAULT = false;
-   public static final int ART_ID_SENTINEL = -1;
+   public static final Long ART_ID_SENTINEL = Long.valueOf(-1);
    public static final String DEFAULT_RATIONALE = "";
    public static final String DEFAULT_NAME = "";
    public static final String DEFAULT_COMMENT = "";
 
-   public static final Integer DEFAULT_ITEM_ID = -1;
+   public static final Long DEFAULT_ITEM_ID = -1L;
    public static final long DEFAULT_TYPE_UUID = -1L;
 
    public static final ModificationType DEFAULT_MODIFICATION_TYPE = ModificationType.NEW;
