+++ plugins/org.eclipse.osee.framework.skynet.core/src/org/eclipse/osee/framework/skynet/core/change/ChangeBuilder.java
@@ -29,7 +29,7 @@ public abstract class ChangeBuilder {
    private final ArtifactTypeId artifactType;
    private final boolean isHistorical;
 
-   public ChangeBuilder(BranchId branch, ArtifactTypeId artifactType, GammaId sourceGamma, int artId, TransactionDelta txDelta, ModificationType modType, boolean isHistorical) {
+   public ChangeBuilder(BranchId branch, ArtifactTypeId artifactType, GammaId sourceGamma, long artId, TransactionDelta txDelta, ModificationType modType, boolean isHistorical) {
       this.sourceGamma = sourceGamma;
       this.artId = ArtifactId.valueOf(artId);
       this.txDelta = txDelta;
