+++ plugins/org.eclipse.osee.framework.skynet.core/src/org/eclipse/osee/framework/skynet/core/relation/RelationLink.java
@@ -43,7 +43,7 @@ import org.eclipse.osee.framework.skynet.core.transaction.SkynetTransaction;
  * @author Ryan D. Brooks
  */
 public class RelationLink implements HasBranch {
-   private int relationId;
+   private Long relationId;
    private GammaId gammaId;
    private String rationale;
    private final RelationTypeToken relationType;
@@ -59,7 +59,7 @@ public class RelationLink implements HasBranch {
 
    private boolean useBackingData;
 
-   public RelationLink(ArtifactToken artifactA, ArtifactToken artifactB, BranchId branch, RelationTypeToken relationType, int relationId, GammaId gammaId, String rationale, ModificationType modificationType, ApplicabilityId applicabilityId) {
+   public RelationLink(ArtifactToken artifactA, ArtifactToken artifactB, BranchId branch, RelationTypeToken relationType, Long relationId, GammaId gammaId, String rationale, ModificationType modificationType, ApplicabilityId applicabilityId) {
       this.relationType = relationType;
       this.relationId = relationId;
       this.gammaId = gammaId;
@@ -300,11 +300,11 @@ public class RelationLink implements HasBranch {
       ArtifactCache.updateCachedArtifact(artifactB);
    }
 
-   public void internalSetRelationId(int relationId) {
+   public void internalSetRelationId(Long relationId) {
       this.relationId = relationId;
    }
 
-   public int getId() {
+   public Long getId() {
       return relationId;
    }
