+++ plugins/org.eclipse.osee.framework.skynet.core/src/org/eclipse/osee/framework/skynet/core/revision/ChangeData.java
@@ -120,7 +120,7 @@ public class ChangeData {
       Set<Artifact> artifacts = new HashSet<>();
       if (kindType == KindType.Artifact || kindType == KindType.ArtifactOrRelation || kindType == KindType.Relation) {
          if (!isEmpty()) {
-            HashMap<Integer, Boolean> excludeArtifact = new HashMap<>();
+            HashMap<Long, Boolean> excludeArtifact = new HashMap<>();
             for (Change change : changes) {
                Artifact artifact = change.getChangeArtifact();
 
@@ -128,7 +128,7 @@ public class ChangeData {
                if (artifactTypesToIgnore != null && !artifactTypesToIgnore.isEmpty()) {
                   try {
                      if (change instanceof AttributeChange) {
-                        Integer id = artifact.getArtId();
+                        Long id = artifact.getArtId();
                         long typeId = change.getItemTypeId().getId();
                         AttributeType attributeType = AttributeTypeManager.getTypeById(typeId);
                         if (excludeArtifact.containsKey(id)) {
@@ -173,7 +173,7 @@ public class ChangeData {
             if (artifactTypesToIgnore != null && !artifactTypesToIgnore.isEmpty()) {
                Set<Artifact> excludeList = new HashSet<>();
                for (Artifact artifactToCheck : artifacts) {
-                  Integer id = artifactToCheck.getArtId();
+                  Long id = artifactToCheck.getArtId();
                   Boolean remove = excludeArtifact.get(id);
                   if (remove != null) {
                      if (remove) {
